+++ cmd/entrypoints/migrate.go
@@ -2,6 +2,7 @@ package entrypoints
 
 import (
 	"context"
+
 	"github.com/flyteorg/flyteadmin/pkg/server"
 
 	"github.com/spf13/cobra"
