+++ pkg/repositories/database.go
@@ -3,10 +3,11 @@ package repositories
 import (
 	"context"
 	"fmt"
-	"gorm.io/driver/sqlite"
 	"io/ioutil"
 	"os"
 	"strings"
+
+	"gorm.io/driver/sqlite"
 
 	runtimeInterfaces "github.com/flyteorg/flyteadmin/pkg/runtime/interfaces"
 	"github.com/flyteorg/flytestdlib/logger"
